requirements.txt
dj-database-url==2.1.0
Django==4.2.6
django-extensions==3.2.3
django-phone-field==1.8.1
django-phonenumber-field==7.2.0
djangorestframework==3.14.0
pylsqpack==0.3.17
pymssql==2.2.2
PyMySQL==1.1.0
